requirements.txt
fastapi
uvicorn
deta
pydantic
cryptography
python-jose